split RouteParamsMiddleware.ts
