+++ app/client/src/ee/RouteParamsMiddleware.ts
@@ -1,0 +1,3 @@
+export * from "ce/RouteParamsMiddleware";
+import { default as CE_RouteParamsMiddleware } from "ce/RouteParamsMiddleware";
+export default CE_RouteParamsMiddleware;
